refactor(rsc): Move Shared API Contracts, Constants, and Clients to hooks package
