+++ apps/tarot/src/api/tarot-api.ts
@@ -1,3 +1,17 @@
+import {
+    TAROT_API_ENV_KEYS,
+    type ReadingCitation,
+    type ReadingHistoryItem,
+    type ReadingHistoryResponse,
+    type ReadingItem,
+    type ReadingPositionResponse,
+    type ReadingRequest,
+    type ReadingResponse,
+    type TarotApiConfig,
+    type TarotApiClient,
+    createTarotApiClient as createSharedTarotApiClient
+} from '@simpletarot/hooks';
+
 export type {
     ReadingCitation,
     ReadingHistoryItem,
@@ -5,34 +19,12 @@ export type {
     ReadingItem,
     ReadingPositionResponse,
     ReadingRequest,
-    ReadingResponse
-} from '@/readings/reading-contracts';
-
-import type {
-    ReadingHistoryResponse,
-    ReadingRequest,
-    ReadingResponse
-} from '@/readings/reading-contracts';
-
-export type TarotApiConfig = {
-    baseUrl: string;
+    ReadingResponse,
+    TarotApiClient,
+    TarotApiConfig
 };
 
-export type TarotApiClient = {
-    createReading: (request: ReadingRequest) => Promise<ReadingResponse>;
-    listReadings: () => Promise<ReadingHistoryResponse>;
-};
-
-type CreateTarotApiClientOptions = TarotApiConfig & {
-    accessToken: string;
-};
-
-type RequestMetadata = {
-    method: 'GET' | 'POST';
-    url: string;
-};
-
-const readRequiredEnv = (key: 'EXPO_PUBLIC_TAROT_API_URL') => {
+const readRequiredEnv = (key: typeof TAROT_API_ENV_KEYS.apiUrl) => {
     const value = process.env[key]?.trim();
 
     if (!value) {
@@ -44,101 +36,10 @@ const readRequiredEnv = (key: 'EXPO_PUBLIC_TAROT_API_URL') => {
 
 const trimTrailingSlashes = (value: string) => value.replace(/\/+$/, '');
 
-const previewBody = (body: string) => body.slice(0, 240);
-
-const contentTypeFor = (response: Response) => response.headers.get('content-type') ?? '';
-
-async function parseJsonResponse<T>(
-    response: Response,
-    request: RequestMetadata
-): Promise<T> {
-    const contentType = contentTypeFor(response);
-    const textBody = await response.text();
-
-    if (!contentType.toLowerCase().includes('application/json')) {
-        console.warn('[tarot-api] non-json response', {
-            bodyPreview: previewBody(textBody),
-            contentType,
-            method: request.method,
-            status: response.status,
-            url: request.url
-        });
-
-        throw new Error(
-            `Tarot API returned ${contentType || 'non-JSON content'} for ${request.method} ${request.url} with status ${response.status}.`
-        );
-    }
-
-    let body: unknown;
-
-    try {
-        body = textBody ? JSON.parse(textBody) : null;
-    } catch (error) {
-        console.warn('[tarot-api] invalid-json response', {
-            bodyPreview: previewBody(textBody),
-            contentType,
-            method: request.method,
-            status: response.status,
-            url: request.url
-        });
-
-        throw error;
-    }
-
-    if (!response.ok) {
-        const message =
-            body &&
-            typeof body === 'object' &&
-            'message' in body &&
-            typeof body.message === 'string'
-                ? body.message
-                : `Request failed with status ${response.status}.`;
-
-        throw new Error(message);
-    }
-
-    return body as T;
-}
-
 export function getTarotApiConfig(): TarotApiConfig {
     return {
-        baseUrl: trimTrailingSlashes(readRequiredEnv('EXPO_PUBLIC_TAROT_API_URL'))
+        baseUrl: trimTrailingSlashes(readRequiredEnv(TAROT_API_ENV_KEYS.apiUrl))
     };
 }
 
-export function createTarotApiClient({
-    accessToken,
-    baseUrl
-}: CreateTarotApiClientOptions): TarotApiClient {
-    const apiBaseUrl = trimTrailingSlashes(baseUrl);
-    const authHeaders = {
-        Authorization: `Bearer ${accessToken}`
-    };
-
-    return {
-        async createReading(request) {
-            const url = `${apiBaseUrl}/readings`;
-            const method = 'POST';
-            const response = await fetch(url, {
-                body: JSON.stringify(request),
-                headers: {
-                    ...authHeaders,
-                    'Content-Type': 'application/json'
-                },
-                method: 'POST'
-            });
-
-            return parseJsonResponse<ReadingResponse>(response, { method, url });
-        },
-        async listReadings() {
-            const url = `${apiBaseUrl}/readings`;
-            const method = 'GET';
-            const response = await fetch(url, {
-                headers: authHeaders,
-                method
-            });
-
-            return parseJsonResponse<ReadingHistoryResponse>(response, { method, url });
-        }
-    };
-}
+export const createTarotApiClient = createSharedTarotApiClient;
