+++ apps/tarot/src/avatars/avatar-api.ts
@@ -1,15 +1,14 @@
-import type { AvatarsResponse } from './avatar-contracts';
+import {
+    TAROT_API_ENV_KEYS,
+    type AvatarApiClient,
+    type AvatarApiConfig,
+    type AvatarsResponse,
+    createAvatarApiClient as createSharedAvatarApiClient
+} from '@simpletarot/hooks';
 
-export type AvatarApiConfig = {
-    accessToken?: string;
-    baseUrl: string;
-};
+export type { AvatarApiClient, AvatarApiConfig, AvatarsResponse };
 
-export type AvatarApiClient = {
-    listAvatarThumbnails: () => Promise<AvatarsResponse>;
-};
-
-const readRequiredEnv = (key: 'EXPO_PUBLIC_TAROT_API_URL') => {
+const readRequiredEnv = (key: typeof TAROT_API_ENV_KEYS.apiUrl) => {
     const value = process.env[key]?.trim();
 
     if (!value) {
@@ -21,94 +20,10 @@ const readRequiredEnv = (key: 'EXPO_PUBLIC_TAROT_API_URL') => {
 
 const trimTrailingSlashes = (value: string) => value.replace(/\/+$/, '');
 
-type RequestMetadata = {
-    method: 'GET';
-    url: string;
-};
-
-const previewBody = (body: string) => body.slice(0, 240);
-
-const contentTypeFor = (response: Response) => response.headers.get('content-type') ?? '';
-
-async function parseJsonResponse<T>(
-    response: Response,
-    request: RequestMetadata
-): Promise<T> {
-    const contentType = contentTypeFor(response);
-    const textBody = await response.text();
-
-    if (!contentType.toLowerCase().includes('application/json')) {
-        console.warn('[avatar-api] non-json response', {
-            bodyPreview: previewBody(textBody),
-            contentType,
-            method: request.method,
-            status: response.status,
-            url: request.url
-        });
-
-        throw new Error(
-            `Avatar API returned ${contentType || 'non-JSON content'} for ${request.method} ${request.url} with status ${response.status}.`
-        );
-    }
-
-    let body: unknown;
-
-    try {
-        body = textBody ? JSON.parse(textBody) : null;
-    } catch (error) {
-        console.warn('[avatar-api] invalid-json response', {
-            bodyPreview: previewBody(textBody),
-            contentType,
-            method: request.method,
-            status: response.status,
-            url: request.url
-        });
-
-        throw error;
-    }
-
-    if (!response.ok) {
-        const message =
-            body &&
-            typeof body === 'object' &&
-            'message' in body &&
-            typeof body.message === 'string'
-                ? body.message
-                : `Request failed with status ${response.status}.`;
-
-        throw new Error(message);
-    }
-
-    return body as T;
-}
-
 export function getAvatarApiConfig(): AvatarApiConfig {
     return {
-        baseUrl: trimTrailingSlashes(readRequiredEnv('EXPO_PUBLIC_TAROT_API_URL'))
+        baseUrl: trimTrailingSlashes(readRequiredEnv(TAROT_API_ENV_KEYS.apiUrl))
     };
 }
 
-export function createAvatarApiClient({
-    accessToken,
-    baseUrl
-}: AvatarApiConfig): AvatarApiClient {
-    const apiBaseUrl = trimTrailingSlashes(baseUrl);
-    const authHeaders = accessToken
-        ? {
-              Authorization: `Bearer ${accessToken}`
-          }
-        : undefined;
-
-    return {
-        async listAvatarThumbnails() {
-            const url = `${apiBaseUrl}/avatars`;
-            const method = 'GET';
-            const response = await fetch(url, {
-                ...(authHeaders ? { headers: authHeaders } : {}),
-                method
-            });
-
-            return parseJsonResponse<AvatarsResponse>(response, { method, url });
-        }
-    };
-}
+export const createAvatarApiClient = createSharedAvatarApiClient;
