+++ apps/tarot/src/readings/server-actions.ts
@@ -2,12 +2,9 @@
 
 import 'server-only';
 
+import { createOneCardReadingRequest } from '@simpletarot/hooks';
 import { createTarotApiClient, getTarotApiConfig } from '@/api/tarot-api';
-import type {
-    ReadingHistoryResponse,
-    ReadingRequest,
-    ReadingResponse
-} from './reading-contracts';
+import type { ReadingHistoryResponse, ReadingResponse } from './reading-contracts';
 
 export type CreateOneCardReadingInput = {
     accessToken: string;
@@ -20,23 +17,6 @@ const createServerClient = (accessToken: string) =>
         accessToken
     });
 
-const oneCardReadingRequest = (question?: string): ReadingRequest => {
-    const trimmedQuestion = question?.trim();
-
-    return {
-        spread: 'single_card',
-        ...(trimmedQuestion ? { question: trimmedQuestion } : {}),
-        items: [
-            {
-                cardIndex: 0,
-                cardName: 'The Fool',
-                position: 'guidance',
-                reversed: false
-            }
-        ]
-    };
-};
-
 export async function listReadingsOnServer(
     accessToken: string
 ): Promise<ReadingHistoryResponse> {
@@ -48,6 +28,6 @@ export async function createOneCardReadingOnServer({
     question
 }: CreateOneCardReadingInput): Promise<ReadingResponse> {
     return createServerClient(accessToken).createReading(
-        oneCardReadingRequest(question)
+        createOneCardReadingRequest(question)
     );
 }
