+++ packages/hooks/src/avatars/avatar-contracts.ts
@@ -1,0 +1,3 @@
+export type AvatarsResponse = {
+    thumbnails: string[];
+};
