+++ packages/hooks/src/constants/tarot-api.ts
@@ -1,0 +1,7 @@
+export const TAROT_API_ENV_KEYS = {
+    apiUrl: 'EXPO_PUBLIC_TAROT_API_URL'
+} as const;
+
+export const READINGS_ENDPOINT_PATH = '/readings';
+export const AVATARS_ENDPOINT_PATH = '/avatars';
+export const JSON_CONTENT_TYPE = 'application/json';
